requirements.txt
mosaicml-composer>=0.16.0
torch>=2.0.0
torchvision>=0.15.0
captum>=0.6.0
optuna>=3.3.0
wandb>=0.15.0
plotly>=5.15.0
dash>=2.11.0
numpy>=1.24.0
pandas>=2.0.0 